LayoutTests/inspector/debugger/snippets-model-expected.txt

Tests snippets model.


Running: testCreateRenameEditDelete
Dumping saved snippets:
Dumping model snippets:
Snippet created.
Dumping saved snippets:
    Snippet: id = 1, name = 'Snippet #1', content = ''.
Dumping model snippets:
    Snippet: id = 1, name = 'Snippet #1', content = ''.
Snippet renamed.
Dumping saved snippets:
    Snippet: id = 1, name = 'New snippet name', content = ''.
Dumping model snippets:
    Snippet: id = 1, name = 'New snippet name', content = ''.
Snippet content changed.
Dumping saved snippets:
    Snippet: id = 1, name = 'New snippet name', content = 'New snippet content'.
Dumping model snippets:
    Snippet: id = 1, name = 'New snippet name', content = 'New snippet content'.
Another snippet created.
Dumping saved snippets:
    Snippet: id = 1, name = 'New snippet name', content = 'New snippet content'.
    Snippet: id = 2, name = 'Snippet #2', content = ''.
Dumping model snippets:
    Snippet: id = 1, name = 'New snippet name', content = 'New snippet content'.
    Snippet: id = 2, name = 'Snippet #2', content = ''.
Snippet deleted.
Dumping saved snippets:
    Snippet: id = 2, name = 'Snippet #2', content = ''.
Dumping model snippets:
    Snippet: id = 2, name = 'Snippet #2', content = ''.
Another snippet deleted.
Dumping saved snippets:
Dumping model snippets:

Running: testEvaluate
Last evaluation source url for snippet: snippets://1_1
Snippet script added, sourceURL = snippets://1_1
Last evaluation source url for snippet: snippets://2_2
Snippet script added, sourceURL = snippets://2_2
Last evaluation source url for snippet: snippets://1_3
Snippet script added, sourceURL = snippets://1_3
